requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.0
django-filter==23.5
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
pytz==2023.3
djangorestframework-simplejwt==5.3.1
